test(electrostatics): 实现3D静电场测试用例
- 添加了3D静电场问题的测试用例
- 实现了从COMSOL网格文件读取和参考数据对比的功能- 优化了边界条件设置方法，提高了代码可读性和效率- 增加了对计算结果的详细验证，包括最大误差、均方根误差等指标- 重构了部分代码结构，提高了可维护性
