+++ fem/kernels/HeatDiffusionKernel.hpp
@@ -14,23 +14,19 @@ namespace FEM {
             Eigen::Matrix<double, TNumNodes, TNumNodes> K_elem;
             K_elem.setZero();
 
-            // FEValues现在需要知道分析类型
+            // FEValues 不再需要知道分析类型来构建B矩阵
             FEValues fe_values(element, 1, AnalysisType::SCALAR_DIFFUSION);
             const MaterialProperty& k_prop = mat_.getProperty("thermal_conductivity");
 
             for (size_t q = 0; q < fe_values.n_quad_points(); ++q) {
                 fe_values.reinit(q);
 
-                // 在一个完整实现中，我们会先插值温度T_q
-
-                // 简化：暂时仍使用常数
                 double k = k_prop.evaluate();
 
-                // 构建B矩阵 (梯度算子)
-                const auto& dN_dx = fe_values.dN_dx();
-                Eigen::MatrixXd B = dN_dx;
-
-                double D = k; // 获取D矩阵 (标量)
+                // --- 正确的实现 ---
+                // 内核(Kernel)现在负责从FEValues获取梯度dN_dx，并将其用作B矩阵
+                const auto& B = fe_values.dN_dx();
+                double D = k; // 材料本构关系 (对于热传导是标量)
 
                 // --- 通用形式： K_elem += B^T * D * B * dV ---
                 K_elem += (B.transpose() * D * B) * fe_values.JxW();
